+++ grass-diary/src/pages/DiaryDetail/Setting.tsx
@@ -15,9 +15,7 @@ import { useModal } from '@state/modal/useModal';
 import { MODAL } from '@constants/message';
 import { INTERACTION } from '@styles/interaction';
 import { useDeleteDiaryDetail } from '@hooks/api/useDeleteDiaryDetail';
-import { usePatchDiary } from '@hooks/api/usePatchDiary';
-import { useToast } from '@state/toast/useToast';
-import { useQueryClient } from '@tanstack/react-query';
+import { usePatchVisibility } from '@hooks/api/usePatchVisibility';
 
 type SettingProps = {
   diaryId: Id;
@@ -28,25 +26,16 @@ type SettingProps = {
 const Setting = ({ diaryId, createdDate, detail }: SettingProps) => {
   const navigate = useNavigate();
   const { modal } = useModal();
-  const { toast } = useToast();
   const { date } = useTodayDate();
   const [canEdit, setCanEdit] = useState(false);
-  const { mutate } = useDeleteDiaryDetail(diaryId);
-  // 일기 patch api는 당일 수정만 가능하도록 되어있음. => 수정해야함.
-  const { mutate: patchDiary } = usePatchDiary(diaryId);
-  const queryClient = useQueryClient();
+  const { mutate: deleteDiary } = useDeleteDiaryDetail(diaryId);
 
-  const editPrivate = () => {
+  const editVisibility = () => {
     if (detail) {
-      const hashArr = detail.tags.map(tag => tag.tag);
-
-      const request = {
-        content: detail.content,
-        isPrivate: detail.isPrivate ? false : true,
-        conditionLevel: `LEVEL_${detail.transparency * 10}`,
-        hashtags: hashArr,
-        imageId: detail.image.length ? detail.image[0].imageId : 0,
-      };
+      const { mutate: patchVisibility } = usePatchVisibility(
+        diaryId,
+        detail?.isPrivate,
+      );
 
       const setting = {
         title: detail.isPrivate ? '일기 공개하기' : '일기 비공개하기',
@@ -63,17 +52,7 @@ const Setting = ({ diaryId, createdDate, detail }: SettingProps) => {
       const button2 = {
         active: true,
         text: detail.isPrivate ? '공개하기' : '비공개하기',
-        clickHandler: () =>
-          patchDiary(request, {
-            onSuccess: () => {
-              toast(
-                detail.isPrivate ? '일기를 공개했어요' : '일기를 비공개했어요 ',
-              );
-              queryClient.invalidateQueries({
-                queryKey: ['get-diaryDetail'],
-              });
-            },
-          }),
+        clickHandler: patchVisibility,
         color: semantic.light.accent.solid.alternative,
         interaction: INTERACTION.accent.subtle(),
       };
@@ -113,7 +92,7 @@ const Setting = ({ diaryId, createdDate, detail }: SettingProps) => {
     const button2 = {
       active: true,
       text: MODAL.delete_diary.button,
-      clickHandler: mutate,
+      clickHandler: deleteDiary,
       color: semantic.light.feedback.solid.negative,
     };
 
@@ -140,7 +119,7 @@ const Setting = ({ diaryId, createdDate, detail }: SettingProps) => {
     <Menus icon={more}>
       {detail && (
         <Menu
-          onClick={editPrivate}
+          onClick={editVisibility}
           text={detail.isPrivate ? '공개하기' : '비공개하기'}
           svg={detail.isPrivate ? lockOpen : lock}
         />
